refactor: move sandwich-ui and storybook to packages
- rename package names
- update references
- move Dockerfiles to package roots
