+++ .github/workflows/react-pull-request.yml
@@ -1,14 +1,12 @@
-name: Storybook Pull Request
+name: "@sandwich-ui/react Pull Request"
 
 on:
   pull_request:
     branches:
       - main
     paths:
-      - ".github/workflows/storybook-pull-request.yml"
-      - "storybook/**"
-      - "storybook.Dockerfile"
-      - "sandwich-ui/**"
+      - ".github/workflows/react-pull-request.yml"
+      - "packages/react/**"
 
 jobs:
   setup:
@@ -37,7 +35,7 @@ jobs:
         uses: ./.github/actions/setup-env
 
       - name: Build
-        run: pnpm build --filter @sandwich-ui/storybook
+        run: pnpm build --filter @sandwich-ui/react
 
   lint:
     needs: setup
@@ -53,26 +51,4 @@ jobs:
         uses: ./.github/actions/setup-env
 
       - name: Lint
-        run: pnpm lint --filter @sandwich-ui/storybook
-
-  docker:
-    runs-on: ubuntu-latest
-    name: Build Docker Image
-
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v3
-
-      - name: Build
-        uses: docker/build-push-action@v5
-        with:
-          context: .
-          push: false
-          file: ./storybook.Dockerfile
-          tags: image:latest
-          cache-from: type=gha
-          cache-to: type=gha,mode=max
+        run: pnpm lint --filter @sandwich-ui/react
