+++ .github/workflows/release.yml
@@ -19,7 +19,7 @@ jobs:
         uses: ./.github/actions/setup-env
 
       - name: Build
-        run: pnpm build --filter @mrsquaare/sandwich-ui
+        run: pnpm build --filter @sandwich-ui/react
 
       - name: Create Release Pull Request or Publish to npm
         uses: changesets/action@v1
