+++ PREVIOUS_CHANGELOGS.md
@@ -1,3 +1,19 @@
+# @mrsquaare/sandwich-ui
+
+## 0.3.1
+
+### Patch Changes
+
+- [#3](https://github.com/MrSquaare/sandwich-ui/pull/3) [`b2f564c`](https://github.com/MrSquaare/sandwich-ui/commit/b2f564cbfdd825519727065e9037e2f6c6d9c02f) Thanks [@MrSquaare](https://github.com/MrSquaare)! - Bump packages, migrate to ESLint v9
+
+## 0.3.0
+
+### Minor Changes
+
+- [#1](https://github.com/MrSquaare/sandwich-ui/pull/1) [`67e57ad`](https://github.com/MrSquaare/sandwich-ui/commit/67e57ad4e631cc20297386ad3099dfe48aceae87) Thanks [@MrSquaare](https://github.com/MrSquaare)! - Initial version
+
+---
+
 # @sandwich-ui/storybook
 
 ## 0.1.1
@@ -7,7 +23,7 @@
 - [#3](https://github.com/MrSquaare/sandwich-ui/pull/3) [`b2f564c`](https://github.com/MrSquaare/sandwich-ui/commit/b2f564cbfdd825519727065e9037e2f6c6d9c02f) Thanks [@MrSquaare](https://github.com/MrSquaare)! - Bump packages, migrate to ESLint v9
 
 - Updated dependencies [[`b2f564c`](https://github.com/MrSquaare/sandwich-ui/commit/b2f564cbfdd825519727065e9037e2f6c6d9c02f)]:
-  - @mrsquaare/sandwich-ui@0.3.1
+  - @sandwich-ui/react@0.3.1
 
 ## 0.1.0
 
@@ -18,4 +34,4 @@
 ### Patch Changes
 
 - Updated dependencies [[`67e57ad`](https://github.com/MrSquaare/sandwich-ui/commit/67e57ad4e631cc20297386ad3099dfe48aceae87)]:
-  - @mrsquaare/sandwich-ui@0.3.0
+  - @sandwich-ui/react@0.3.0
