+++ docs/src/content/docs/guide/getting-started.mdx
@@ -10,17 +10,17 @@ import { Tabs, TabItem } from '@astrojs/starlight/components';
 <Tabs>
 <TabItem label="npm">
 ```bash
-npm install @mrsquaare/sandwich-ui
+npm install @sandwich-ui/react
 ```
 </TabItem>
 <TabItem  label="yarn">
 ```bash
-yarn add @mrsquaare/sandwich-ui
+yarn add @sandwich-ui/react
 ```
 </TabItem>
 <TabItem  label="pnpm">
 ```bash
-pnpm install @mrsquaare/sandwich-ui
+pnpm install @sandwich-ui/react
 ```
 </TabItem>
 </Tabs>
