+++ packages/react-storybook/package.json
@@ -1,6 +1,6 @@
 {
   "private": true,
-  "name": "@sandwich-ui/storybook",
+  "name": "@sandwich-ui/react-storybook",
   "version": "0.1.1",
   "type": "module",
   "scripts": {
@@ -13,7 +13,7 @@
     "clean": "rimraf storybook-static/ node_modules/.cache/"
   },
   "dependencies": {
-    "@mrsquaare/sandwich-ui": "workspace:*",
+    "@sandwich-ui/react": "workspace:*",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
