+++ packages/react-storybook/src/components/Card.stories.tsx
@@ -5,7 +5,7 @@ import {
   CardFooter,
   CardHeader,
   CardTitle,
-} from "@mrsquaare/sandwich-ui";
+} from "@sandwich-ui/react";
 import type { Meta, StoryObj } from "@storybook/react";
 
 const meta = {
