+++ packages/react-storybook/src/components/Navigation.stories.tsx
@@ -10,7 +10,7 @@ import {
   Navigation,
   NavigationItem,
   NavigationList,
-} from "@mrsquaare/sandwich-ui";
+} from "@sandwich-ui/react";
 import type { Meta, StoryObj } from "@storybook/react";
 
 const meta = {
