+++ packages/react-storybook/src/components/Tooltip.stories.tsx
@@ -2,7 +2,7 @@ import {
   Tooltip,
   TooltipContent,
   TooltipTrigger,
-} from "@mrsquaare/sandwich-ui";
+} from "@sandwich-ui/react";
 import type { Meta, StoryObj } from "@storybook/react";
 
 const meta = {
