+++ packages/react/src/components/Avatar.tsx
