+++ packages/react/tsconfig.json
