+++ packages/react/vite.config.ts
