--- sandwich-ui/CHANGELOG.md
@@ -1,13 +1,0 @@
-# @mrsquaare/sandwich-ui
-
-## 0.3.1
-
-### Patch Changes
-
-- [#3](https://github.com/MrSquaare/sandwich-ui/pull/3) [`b2f564c`](https://github.com/MrSquaare/sandwich-ui/commit/b2f564cbfdd825519727065e9037e2f6c6d9c02f) Thanks [@MrSquaare](https://github.com/MrSquaare)! - Bump packages, migrate to ESLint v9
-
-## 0.3.0
-
-### Minor Changes
-
-- [#1](https://github.com/MrSquaare/sandwich-ui/pull/1) [`67e57ad`](https://github.com/MrSquaare/sandwich-ui/commit/67e57ad4e631cc20297386ad3099dfe48aceae87) Thanks [@MrSquaare](https://github.com/MrSquaare)! - Initial version
